Example of cross-cutting authentication/authorization example
This is an example of a way to compose Akka gRPC with 'regular' Akka HTTP code
to achieve server-side authentication/authorization when it is a true
'cross-cutting concern': the implementation is still based on the regular API
trait (and does not need to know about the token), and is wrapped in a
Directive that performs the token verification.

This is not a complete example: documentation, a Java version and test coverage
are still missing, and authentication errors are reported as a 'normal' HTTP
400 response rather than a gRPC error code 7. Nonetheless it might be a good
starting point for further exploration.