Fix #636 by publishing bat scripts for Windows (#644)
The published `bat` scripts are generated by copying the assembly JARs
generated by sbt-assembly, but after having skipped the bash specific
part in the beginning. This is somewhat problematic, because we are
relying on the internal format of the JARs generated by sbt-assembly.
A more solid approach would just generate two assemblies, but with
different settings. Unfortunately, configuring the build to do this is
all but straight forward.