play nice on projects with other sbt-protoc-based plugins
When AkkaGrpcPlugin is added to a project with other plugins
appending generators to PB.targets, these extra generators could
be ignored, when AkkaGrpcPlugin was injected last (something which
cannot be controlled, unless there is an explicit dependency from
that plugin to AkkaGrpcPlugin via `AutoPlugin#requires`).

I could not find in the original commit 7f2f96c a reason for
overriding PB.targets, so akka-grpc generators are now appended.
Writing a scripted test to guard this behavior is hard as the order
in which plugins are flattened is not guaranteed/deterministic.