Plug Akka Discovery into the grpc-java NameResolver
Instead of performing discovery 'outside' of the Channel.

This would allow configuring client-side loadbalancing, #809, but I have
not yet tested that that goal is actually achieved.

Before un-drafting this generally needs some more testing whether it
does the right thing, especially re-resolving names in case of updates
etc. It may also mean we can simplify/remove the recreation code in
ClientState.scala, but that also still needs to be verified.