Avoid deadlock (#1178)
* Avoid deadlock

The 'concat' is used to make sure the stream fails with the right error
when the trailing headers contain an error status code. To avoid
deadlock we should put that after the watchTermination that successfully
terminates the stream where there are no trailing headers at all.

* Update Akka HTTP

To fix the remaining timeouts caused by https://github.com/akka/akka-http/pull/3619