Add test for broken retries with netty backend
withConnectionAttempts() does not work as expected with the Netty backend:
if a request is sent before the server is started, the caller sees a
failure immediately and no retries are attempted.

with the Akka-Http backend, the same test works as expected, i.e. one of
the retries will reach the server once it's up and the caller will see
a succes result.