perf: Trailers only support (#1973)
An optimisation exists in the gRPC spec that allows responses that immediately
return an error to return it as a "Trailers-Only" response. This is where the
response has no body, rather, the `grpc-status`, `grpc-message` and
`grpc-status-details-bin` trailers are placed in the response headers frame.

When adding support for this, I found a bug in the Akka gRPC client where
trailers only responses were not correctly handled in streamed requests
(the error returned to the client code was an
`AbruptStageTerminationException`).