Don't send content length with gRPC responses
This avoids sending a content length header with gRPC unary responses.

The gRPC spec specifies require a content length to be sent, though it also
doesn't expressly forbid it. It seems server implementations probably don't
usually send one. There are some clients that don't expect one, most notably,
the envoy gRPC web transcoder doesn't expect one, and fails to strip it out
before modifying the response body, causing protocol errors:

https://github.com/envoyproxy/envoy/issues/40097

One downside of this change is that it is probably a performance regression
for unary requests, as sending a `Strict` entity I believe is more performant
than sending a `Chunked` entity - no substream needs to be materialised when
sending a strict entity. I'm not sure exactly how much more performant it is
though, and the difference is likely negligable for services that actually do
something like access a database.